// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/mktreferral/Shared/DTOs/patient-referral-commission.dto.ts
export class PatientVisitLevelReferralCommission_DTO {
    PatientReferralCommissionId: number = 0;
    VisitDate: string = "";
    PatientId: number = 0;
    PatientVisitId: number = 0;
    ReferringPartyId: number = 0;
    ReferralSchemeId: number = 0;
    ReferralAmount: number = 0;
    Remarks: string = "";
    IsActive: boolean = false;
}